Add support for SMPL-X 2020 model

+++ build/make_addon_zip.sh
@@ -1,25 +1,45 @@
 #!/bin/bash
+
+#BUILD_SMPLX_10=1
+BUILD_SMPLX_300=1
+#BUILD_SMPLX_2020=1
+
 pushd ../..
 
 filedate=$(date '+%Y%m%d')
 
-# Build 10 shape model add-on
-#echo "Generating $archivename"
-#archivename=./smplx_blender_addon_$filedate.zip
-#if [ -f $archivename ]; then
-#  echo "Removing old add-on: $archivename"
-#  rm $archivename
-#fi
-#zip $archivename smplx_blender_addon/*.py smplx_blender_addon/*.md smplx_blender_addon/data/*.npz smplx_blender_addon/data/*.json smplx_blender_addon/data/*.png smplx_blender_addon/data/smplx_model_20210421.blend
-
-# Build 300 shape model add-on
-echo "Generating $archivename"
-archivename=./smplx_blender_addon_300_$filedate.zip
-if [ -f $archivename ]; then
-  echo "Removing old add-on: $archivename"
-  rm $archivename
+if [ -n "$BUILD_SMPLX_10" ]; then
+  # Build 10 shape model add-on
+  archivename=./smplx_blender_addon_$filedate.zip
+  echo "Generating $archivename"
+  if [ -f $archivename ]; then
+    echo "Removing old add-on: $archivename"
+    rm $archivename
+  fi
+  zip $archivename smplx_blender_addon/*.py smplx_blender_addon/*.md smplx_blender_addon/data/*.npz smplx_blender_addon/data/*.json smplx_blender_addon/data/*.png smplx_blender_addon/data/smplx_model_20210421.blend
 fi
-# --compression-method store
-zip $archivename smplx_blender_addon/*.py smplx_blender_addon/*.md smplx_blender_addon/data/*.npz smplx_blender_addon/data/*.json smplx_blender_addon/data/*.png smplx_blender_addon/data/smplx_model_300_20220615.blend smplx_blender_addon/data/smplx_model_lh_300_20230214.blend
+
+if [ -n "$BUILD_SMPLX_300" ]; then
+  # Build 300 shape model add-on
+  archivename=./smplx_blender_addon_300_$filedate.zip
+  echo "Generating $archivename"
+  if [ -f $archivename ]; then
+    echo "Removing old add-on: $archivename"
+    rm $archivename
+  fi
+  # --compression-method store
+  zip $archivename smplx_blender_addon/*.py smplx_blender_addon/*.md smplx_blender_addon/data/*.npz smplx_blender_addon/data/*.json smplx_blender_addon/data/*.png smplx_blender_addon/data/smplx_model_300_20220615.blend smplx_blender_addon/data/smplx_model_lh_300_20230214.blend
+fi
+
+if [ -n "$BUILD_SMPLX_2020" ]; then
+  # Build SMPL-X 2020 model add-on
+  archivename=./smplx_blender_addon_2020_300_100_$filedate.zip
+  echo "Generating $archivename"
+  if [ -f $archivename ]; then
+    echo "Removing old add-on: $archivename"
+    rm $archivename
+  fi
+  zip $archivename smplx_blender_addon/*.py smplx_blender_addon/*.md smplx_blender_addon/data/*.npz smplx_blender_addon/data/*.json smplx_blender_addon/data/*.png smplx_blender_addon/data/smplx_model_2020_300_100_20230227.blend
+fi
 
 popd
